pool: Fix potential integer overflow issue
Flagged by findbugs as a high priority issue.

Target: trunk
Target: 1.9.4
Target: 1.9.3
Target: 1.9.2
Patch: http://rb.dcache.org/r/407/
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>

+++ modules/dCache/org/dcache/pool/movers/XrootdMoverListener.java
@@ -327,13 +327,13 @@ public class XrootdMoverListener implements StreamListener {
 
             if (newFilePointer > currentFileSize) {
 
-                int usedBlocks = (int) (currentFileSize / BLOCK_SIZE);
+                long usedBlocks = (currentFileSize / BLOCK_SIZE);
 
                 if (currentFileSize % BLOCK_SIZE != 0) {
                     usedBlocks++;
                 }
 
-                int expectedBlocks = (int) (newFilePointer / BLOCK_SIZE);
+                long expectedBlocks = (newFilePointer / BLOCK_SIZE);
 
                 if (newFilePointer % BLOCK_SIZE != 0) {
                     expectedBlocks++;
